+++ www/install/convertConfig.php
@@ -1,0 +1,123 @@
+<?php
+
+use go\core\App;
+use go\core\util\IniFile;
+
+require(__DIR__ . "/../vendor/autoload.php");
+
+header('Content-Type: text/plain');	
+
+$configFile = App::findConfigFile('config.php');
+$iniFile = substr($configFile, 0, -3).'ini';
+
+if(file_exists($iniFile)) {	
+	echo "INI file already exists so skipping conversion";
+}
+
+$globalConfig = [];
+if (file_exists('/etc/groupoffice/globalconfig.inc.php')) {
+	require('/etc/groupoffice/globalconfig.inc.php');
+	$globalConfig = $config;
+}
+
+
+
+require($configFile);
+
+$config = array_merge($globalConfig, $config);
+
+
+$values = [
+		'title' => 'title',
+		'language' => 'language',
+		'webmaster_email' => 'systemEmail',
+		'smtp_host' => 'smtpHost',
+		'smtp_port' => 'smtpPort',
+		'smtp_username' => 'smtpUsername',
+		'smtp_password' => 'smtpPassword',
+		'smtp_encryption' => 'smtpEncryption',
+		'password_min_length' => 'passwordMinLength'
+];
+
+foreach ($values as $old => $new) {
+	if (empty($config[$old])) {
+		continue;
+	}
+	$sql = "replace into core_setting select id as moduleId, '" . $new . "' as name, :value as value from core_module where name='core'";
+	$stmt = GO()->getDbConnection()->getPDO()->prepare($sql);
+	$stmt->bindValue(":value", $config[$old]);
+	$stmt->execute();
+
+}
+
+
+
+$values = [
+		'default_timezone' => 'defaultTimezone',
+		'default_time_format' => 'defaultTimeFormat',
+		'default_currency' => 'defaultCurrency',
+		'default_first_weekday' => 'defaultFirstWeekday',
+		'default_list_separator' => 'defaultListSeparator',
+		'default_text_separator' => 'defaultTextSeparator',
+		'default_thousands_separator' => 'defaultThousandSeparator',
+		'default_decimal_separator' => 'defaultDecimalSeparator',
+		'register_user_groups' => 'defaultGroups'
+];
+
+foreach ($values as $old => $new) {
+	if (empty($config[$old])) {
+		continue;
+	}
+	$sql = "replace into core_setting select id as moduleId, '" . $new . "' as name, :value as value from core_module where name='users'";
+	$stmt = GO()->getDbConnection()->getPDO()->prepare($sql);
+	$stmt->bindValue(":value", $config[$old]);
+	$stmt->execute();
+}
+
+if(isset($config['default_date_format']) && isset($config['default_date_separator'])) {
+	$f = $config['default_date_format'][0].
+						$config['default_date_separator'].
+						$config['default_date_format'][1].
+						$config['default_date_separator'].
+						$config['default_date_format'][2];
+	
+	$sql = "replace into core_setting select id as moduleId, 'defaultDateFormat' as name, :value as value from core_module where name='users'";	
+	$stmt = GO()->getDbConnection()->getPDO()->prepare($sql);
+	$stmt->bindValue(":value", $f);
+	$stmt->execute();
+	
+
+}
+
+
+$iniData =  [
+						"general" => [
+								"dataPath" => $config['file_storage_path'] ?? '/home/groupoffice',
+								"tmpPath" => $config['tmpdir'] ?? sys_get_temp_dir() . '/groupoffice',
+								"debug" => !empty($config['debug'])
+						],
+						"db" => [
+								"dsn" => 'mysql:host=' . ($config['db_host'] ?? "localhost"). ';dbname=' . ($config['db_name'] ?? "groupoffice"),
+								"username" => $config['db_user'] ?? "groupoffice",
+								"password" => $config['db_pass'] ?? ""
+						], 
+						"limits" => [
+								"maxUsers" => $config['max_users'] ?? 0,
+								"storageQuota" => $config['quota'] ?? 0,
+								"allowedModules" => $config['allowed_modules'] ?? ""
+						]
+				];
+
+$file = new IniFile();
+$file->create($iniData);
+
+
+if(!is_writable($iniFile)) {
+	echo "Can't write to INI file ". $iniFile.". Please create it with the following content and rerun the upgrade: \n\n";
+	$file->update(['db' => ['password' => '[YOURPASSWORDHERE]']]);
+	echo (string) $file;
+	exit();
+} else
+{
+	$file->write($iniFile);
+}