+++ CHANGELOG.md
@@ -31,6 +31,9 @@ Comments
 - Always use read more links and removed settings
 - Remove comments from edit dialogs. Always use detail views.
 
+ActiveSync
+- Spam folder syncs too now.
+
 
 15-03-2018 6.2.86
 - Moved disk cache directory to file_storage_path because tmp is containerized on newer linux versions.